requirements.txt
# 基础依赖
numpy==1.23.5
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1

# 图像处理
opencv-python==4.12.0.88
Pillow==12.0.0
imagecodecs==2023.9.18

# 3D处理
open3d==0.19.0
trimesh==4.8.3
plyfile==1.1.2
pyquaternion==0.9.9

# 科学计算
scipy==1.10.1
scikit-learn==1.1.1
pandas==2.3.3

# 可视化
matplotlib==3.10.7
seaborn==0.13.2
plotly==6.4.0

# GUI
PyQt5==5.15.11
pyqtgraph==0.13.7
gradio==3.35.2

# 分割模型
segment-anything==1.0
groundingdino==0.1.0
clip @ git+https://github.com/openai/CLIP.git

# 工具库
tqdm==4.67.1
wandb==0.22.2
PyYAML==6.0.3
omegaconf==2.3.0
hydra-core==1.3.2
configargparse==1.7.1

# Web
fastapi==0.119.0
uvicorn==0.38.0
gradio-client==1.13.3

# 其他
supervision==0.26.1
timm==1.0.20
transformers==4.21.0